First attempt at /gems/<gem_name>
We try to generate a graph for classes that appear when the gem is
required that aren't available in a vanilla Ruby environment.

Right now doesn't work 100% since that gem can require other gems that
add more classes to the environment, but that aren't part of itself, and
so aren't very interesting for the user to look at.